+++ packages/expo/CHANGELOG.md
@@ -25,6 +25,7 @@
 - [apple] Migrate remaining `expo-module.config.json` to unified platform syntax. ([#34445](https://github.com/expo/expo/pull/34445) by [@reichhartd](https://github.com/reichhartd))
 - [apple] Move `AppDelegate` integration from `expo-modules-core` to `expo` package. ([#34985](https://github.com/expo/expo/pull/34985) by [@lukmccall](https://github.com/lukmccall))
 - [apple] Add EXAppDelegateWrapper import to Expo.h ([#35172](https://github.com/expo/expo/pull/35172) by [@gabrieldonadel](https://github.com/gabrieldonadel))
+- Drop `fbemitter` in favor of custom, API compatible `DevToolsPluginEventEmitter` event emitter and deprecate `DevToolsPluginClient.eventEmitter`. ([#35318](https://github.com/expo/expo/pull/35318) by [@kitten](https://github.com/kitten))
 
 ## 52.0.37 - 2025-02-20